Move quit confirmation presenter tests
