+++ cmuxTests/ShortcutAndCommandPaletteTests.swift
@@ -1903,100 +1903,6 @@ final class QuitConfirmationPolicyTests: XCTestCase {
     }
 }
 
-@MainActor
-final class QuitConfirmationAlertPresenterTests: XCTestCase {
-    func testPendingTerminateReplyWaitsOnlyForTerminateOwnedConfirmation() {
-        XCTAssertEqual(
-            AppDelegate.pendingTerminateReply(
-                isAwaitingTerminateKills: true,
-                hasActiveQuitConfirmation: false,
-                activeQuitConfirmationOwnsTerminateRequest: false
-            ),
-            .terminateLater
-        )
-        XCTAssertEqual(
-            AppDelegate.pendingTerminateReply(
-                isAwaitingTerminateKills: false,
-                hasActiveQuitConfirmation: true,
-                activeQuitConfirmationOwnsTerminateRequest: true
-            ),
-            .terminateLater
-        )
-        XCTAssertEqual(
-            AppDelegate.pendingTerminateReply(
-                isAwaitingTerminateKills: false,
-                hasActiveQuitConfirmation: true,
-                activeQuitConfirmationOwnsTerminateRequest: false
-            ),
-            .terminateCancel
-        )
-        XCTAssertNil(
-            AppDelegate.pendingTerminateReply(
-                isAwaitingTerminateKills: false,
-                hasActiveQuitConfirmation: false,
-                activeQuitConfirmationOwnsTerminateRequest: false
-            )
-        )
-    }
-
-    func testPresenterUsesSheetCompletionWithoutRunningNestedModalLoop() {
-        let alert = QuitConfirmationAlertSpy()
-        let hostWindow = NSWindow(
-            contentRect: NSRect(x: 0, y: 0, width: 480, height: 320),
-            styleMask: [.titled],
-            backing: .buffered,
-            defer: false
-        )
-
-        var completedResponse: NSApplication.ModalResponse?
-        var completedSuppressionState: NSControl.StateValue?
-        let presenter = QuitConfirmationAlertPresenter(
-            alert: alert,
-            presentingWindowProvider: { hostWindow }
-        ) { response, suppressionState in
-            completedResponse = response
-            completedSuppressionState = suppressionState
-        }
-
-        presenter.present()
-
-        XCTAssertTrue(alert.didBeginSheetModal)
-        XCTAssertFalse(alert.didRunModal)
-        XCTAssertNil(completedResponse)
-
-        alert.capturedSheetCompletion?(.alertFirstButtonReturn)
-
-        XCTAssertEqual(completedResponse, .alertFirstButtonReturn)
-        XCTAssertEqual(completedSuppressionState, .off)
-    }
-}
-
-private final class QuitConfirmationAlertSpy: NSAlert {
-    var didBeginSheetModal = false
-    var didRunModal = false
-    var capturedSheetCompletion: ((NSApplication.ModalResponse) -> Void)?
-
-    override init() {
-        super.init()
-        addButton(withTitle: "Quit")
-        addButton(withTitle: "Cancel")
-        showsSuppressionButton = true
-    }
-
-    override func beginSheetModal(
-        for sheetWindow: NSWindow,
-        completionHandler handler: ((NSApplication.ModalResponse) -> Void)? = nil
-    ) {
-        didBeginSheetModal = true
-        capturedSheetCompletion = handler
-    }
-
-    override func runModal() -> NSApplication.ModalResponse {
-        didRunModal = true
-        return .alertSecondButtonReturn
-    }
-}
-
 
 final class UpdateChannelSettingsTests: XCTestCase {
     func testResolvedFeedFallsBackWhenInfoFeedMissing() {
